feature@add markdown/txt/pdf parser
