+++ data/中国科学院计算技术研究所学生出国（境）管理办法.txt
@@ -1,0 +1,35 @@
+中国科学院计算技术研究所学生出国（境）管理办法
+为保证我所研究生培养工作的正常进行，规范在读研究生出国（境）管理，根据《教育部关于简化大专以上学历人员自费出国留学审批手续的通知》（教外留〔2003〕1号）以及其他相关文件精神，特制定本管理办法。
+一、	在读研究生公派出国（境）
+（一）公派出国（境）包括：政府间交流项目或根据所里的安排出国（境）参加国际学术会议、合作研究、访问考察、中外联合培养、培训、比赛等对外交流活动。
+（二）公派出国（境）审批程序：
+1、填写《中科院计算所出国（境）任务审批书-学生用》（附件1；以下简称“审批书”）；
+2、将填写好的《审批书》送交导师、科研实体负责人签署意见；
+3、持《审批书》及以下材料到研究生部审核签字（或盖章）：
+	出国参加学术会议的，提交会议邀请函并附上大会论文复印件；
+	出国进行合作研究的，提交我所与合作单位双方签署的有效协议等文件；
+	出国中外联合培养的，提交本人与对方签订的“联合培养研究生协议书”。
+4、审核通过后，领取户籍证明、学籍证明、收入证明等相关文件；
+5、学生出国（境）一律持因私证件。特殊情况随同员工一起出访的，可作为陪同人员申请因公证件。
+（三）定向、委培研究生在读期间申请出国（境）参加学术交流活动，应回本人人事关系所在单位办理，涉及的费用由定向、委培单位支付。同时，原单位须向我所研究生部出具同意其出访的公函。
+（四）公派出国(境)研究生，未经所里批准不得延长在外期限。逾期两周不归者，研究生部将依照我所学籍管理规定处理。对于逾期未归的研究生，由导师协同研究生部通知学生本人，为其办理离所手续，将户口、档案转回本人生源地。
+由国家留学基金委或中国科学院等上级主管部门资助的出访项目，参照上述管理规定，按项目具体要求办理。
+二、	在读研究生自费出国（境）探亲或旅游
+研究生在学期间，不得请假出国（境）探亲、旅游，只能利用国家规定的节假日进行。
+（一）自费出国（境）探亲或旅游的审批程序：
+1、填写《中科院计算所研究生申请出国（境）审批表》（附件2；以下简称“审批表”），及《短期因私出国（境）安全承诺书》（附件3；以下简称“承诺书”）；
+2、将《审批表》和《承诺书》交研究生部审核，同时提交以下材料：
+1）探亲：直系亲属的邀请信；
+2）旅游：国家正规旅行社相关材料。
+3、研究生部审核通过后，开具《集体户口卡借用介绍信》。学生持介绍信到国科大学生处户籍管理办公室领取户籍证明。
+4、定向和延期毕业的研究生，申请自费出国（境）旅游，本所不予受理。
+(二) 出国（境）探亲或旅游的研究生，应按期返回。逾期不归者，研究生部将依照我所学生管理相关规定进行处理。
+三、	应届毕业研究生自费出国（境）留学
+应届毕业生自费出国（境）留学或工作的管理办法详见《中科院计算所研究生就业管理暂行办法》。
+本所不受理非应届毕业生在学期间自费出国留学的申请。
+四、 研究生出国（境）期间，一切安全责任自负。
+五、其他
+（一）出国（境）前，将《中国科学院大学学生请假单》交研究生部备案；
+（二）出国（境）后，按时返回的研究生，应及时到研究生部办理销假手续。
+（三）涉密人员须按照所内《计算所涉密人员管理规定》办理，不在本管理办法管辖范围内。
+（四）本管理办法的解释权归研究生部，自公布之日（2013年９月１日起）起执行。原《关于在读研究生出国（境）的管理办法》（2010年11月23日版）同时废止。
